reorganized imports to clean things up a bit
