+++ src/lib/examples/index.js
@@ -1,0 +1,16 @@
+import { context_argument_map } from './examples'
+import { by_book_order, derive_filters } from './filters'
+import ExampleSummary from './ExampleSummary.svelte'
+import Filters from './Filters.svelte'
+import SourceData from './SourceData.svelte'
+import TargetData from './TargetData.svelte'
+
+export {
+	by_book_order,
+	context_argument_map,
+	derive_filters,
+	ExampleSummary,
+	Filters,
+	SourceData,
+	TargetData,
+}
